pyproject.toml
[tool.poetry]
name = "keep-the-lads-in-check"
version = "0.2.0"
description = "A discord bot for my discord server."
authors = ["Arpit Chakladar <54011232+arpitchakladar@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.1"
youtube-dl = "^2021.5.16"
aiohttp = "^3.7.4"
discord = "^1.0.1"
motor = "^2.4.0"
profanityfilter = "^2.0.6"
pymongo = "^3.11.4"
dnspython = "^2.1.0"
PyNaCl = "^1.4.0"
flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
